Added resource_view policy; added thumbnail recreate option
git-svn-id: https://svn.l3s.uni-hannover.de/learnweb/trunk/LearnwebOER/Learnweb@4938 422a2c9f-66af-4668-9993-60dc3b0871b8

+++ src/de/l3s/learnweb/FileManager.java
@@ -244,13 +244,16 @@ public class FileManager
 
         cache.remove(file.getId());
 
-        try
+        if(file.exists())
         {
-            file.getActualFile().delete();
-        }
-        catch(Throwable e)
-        {
-            log.error("Could not delete file: " + file.getId(), e);
+            try
+            {
+                file.getActualFile().delete();
+            }
+            catch(Throwable e)
+            {
+                log.error("Could not delete file: " + file.getId(), e);
+            }
         }
     }
